Replace rigorous definition in .cabal w/ handwave.
While it makes sense to give the rigorous definitions of (co)free
(co)monads in the documentation, the .cabal file (which contains a
high-level summary for the use of programmers trying to decide whether
or not the package is relevant to their problem) is IMHO not the place
for them. If the reader is well-versed in category theory and free
constructions, they'll say "well, yes, it's a free thing defined in the
usual way; I could work that out from the name." If they're unfamiliar
with category theory, they'll be faced with a wall of unfamiliar
terminology and left none the wiser.

This commit replaces the rigorous definitions with non-rigorous
descriptions that are hopefully more comprehensible to the
non-mathematical reader, and adds links to some of @ekmett's blog posts
on the subject.