make `free` forward`-Wcompat`ible (again)
This is follow-up to c9430031460a7c89038dc5e24796b69f3568b61e
but this time we have warning flags we can put in place to get
alerted of regressions.

This represents a breaking change as it changes the class constraints
for some `Monoid` instances.

Note: the constraints on some `MonadFail` could possibly be relaxed